Corrected the antifaro method
Now it correctly leaves the Ace of Spades on the bottom of the deck and also leaves the Ace of Clubs on Top.

Additionally, you can now add a number of times to the shuffle much like the outfaro method. (which leaves the ace of spades on the bottom as well.)